pyproject.toml
[tool.poetry]
name = "mvp_chat"
package-mode = false
version = "0.1.0"
description = "Local LLM Chatbot MVP with FastAPI backend and Streamlit frontend"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
transformers = "^4.41.2"
accelerate = "^0.30.1"
streamlit = "^1.35.0"
openai = "^2.21.0"
python-dotenv = "^1.1.1"
google-auth = "^2.47.0"
langgraph = "^1.0.9"
langchain = "^1.2.10"
langchain-openai = "^1.1.10"
langchain-community = "^0.4.1"
langchain-chroma = "^1.1.0"
sentence-transformers = "^5.2.3"
langchain-huggingface = "^1.2.0"
pymupdf = "^1.27.1"
fpdf2 = "^2.8.6"
python-multipart = "^0.0.22"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
